AssemblyProcessor: further cleaning/refactor
